// generated/src/app/api-models/ipa-address.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class IpaAddress {
  seqIpaAddress: number;
  ipaId: string;
  name2: string;
  addressLine1: string;
  addressLine2: string;
  primaryAddress: string;
  addrCategoryCode: string;
  city: string;
  state: string;
  zipCode: string;
  county: string;
  country: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}